+++ bin/old-cli/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "opup"
+version = "0.1.0"
+edition = "2021"
+authors = ["merklefruit", "refcell"]
+
+[dependencies]
+eyre = "0.6.8"
+inquire = "0.6.2"
+serde = "1.0.163"
+serde_json = "1.0.96"
