+++ bin/old-cli/src/set_addresses.rs
