+++ bin/old-cli/src/stack/challenger_agent.rs
