+++ bin/opup/src/lib.rs
@@ -1,0 +1,8 @@
+/// The CLI entrypoint for the binary.
+pub mod cli;
+
+/// Module containing stack packages.
+pub mod stack;
+
+/// Binary Telemetry
+pub mod telemetry;
